refactor: [#70] ResponseCodeAndMessage -> ResponseCodeAndMessages로 이름 변경
