+++ BE/src/main/java/dev/whatevernote/be/common/BaseResponse.java
@@ -12,7 +12,7 @@ public class BaseResponse<T> {
 		this.data = data;
 	}
 
-	public BaseResponse(ResponseCodeAndMessage codeAndMessage, T data) {
+	public BaseResponse(ResponseCodeAndMessages codeAndMessage, T data) {
 		this.code = codeAndMessage.getCode();
 		this.message = codeAndMessage.getMessage();
 		this.data = data;
