--- BE/src/main/java/dev/whatevernote/be/response/CodeAndMessages.java
@@ -1,8 +1,0 @@
-package dev.whatevernote.be.response;
-
-public interface CodeAndMessages {
-
-	String getCode();
-
-	String getMessage();
-}
