+++ BE/src/main/java/dev/whatevernote/be/web/controller/NoteController.java
@@ -1,10 +1,10 @@
 package dev.whatevernote.be.web.controller;
 
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.NOTE_CREATE_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.NOTE_MODIFY_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.NOTE_REMOVE_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.NOTE_RETRIEVE_ALL_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.NOTE_RETRIEVE_DETAIL_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.NOTE_CREATE_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.NOTE_MODIFY_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.NOTE_REMOVE_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.NOTE_RETRIEVE_ALL_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.NOTE_RETRIEVE_DETAIL_SUCCESS;
 
 import dev.whatevernote.be.common.BaseResponse;
 import dev.whatevernote.be.service.NoteService;
