+++ BE/src/test/java/dev/whatevernote/be/web/controller/ContentControllerTest.java
@@ -1,9 +1,9 @@
 package dev.whatevernote.be.web.controller;
 
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.CONTENT_CREATE_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.CONTENT_MODIFY_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.CONTENT_REMOVE_SUCCESS;
-import static dev.whatevernote.be.common.ResponseCodeAndMessage.CONTENT_RETRIEVE_ALL_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.CONTENT_CREATE_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.CONTENT_MODIFY_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.CONTENT_REMOVE_SUCCESS;
+import static dev.whatevernote.be.common.ResponseCodeAndMessages.CONTENT_RETRIEVE_ALL_SUCCESS;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.refEq;
 import static org.mockito.Mockito.doNothing;
